requirements.txt
# Core
Django>=4.2.0,<5.0.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9

# REST API
djangorestframework>=3.14.0

# Forms & Frontend
django-crispy-forms>=2.0
crispy-bootstrap4>=2022.1

# AI & OpenAI
openai>=1.3.0
tiktoken>=0.5.1

# Image Processing
Pillow>=10.1.0

# Development & Debug
django-debug-toolbar>=4.2.0
black>=23.11.0
flake8>=6.1.0

# Security
django-cors-headers>=4.3.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
